canastarosa/requirements.txt
django==3.0.3
djangorestframework==3.11.0
django-filter==2.2.0
psycopg2-binary==2.8.3
